# swagger-docs.yaml
openapi: 3.0.3
info:
  title: Moneymaze API
  version: '1.0.0'
  description: |
    ## Overview
    RESTful API for personal expense tracking and budget management. Built with NestJS, Prisma, and PostgreSQL.

    ## Features
    - **Authentication**: JWT-based auth with HTTP-only cookies for refresh tokens
    - **User Management**: Profile management with avatar uploads
    - **Categories**: Custom expense categories with icons and colors
    - **Budget Management**: Template-based budgets with automatic period generation
    - **Expense Tracking**: Detailed expense records with category support
    - **Subscriptions**: Recurring subscription management
    - **Analytics**: Expense summaries and spending insights
    - **Health Monitoring**: Comprehensive health checks for production readiness

    ## Response Format
    All API responses follow a standardized format:
    ```json
    {
      "meta": {
        "code": "200000",
        "type": "SUCCESS",
        "message": "Operation completed successfully",
        "timestamp": "2025-11-15T10:30:00.000Z",
        "request_id": "uuid",
        "request_duration": 123
      },
      "data": { /* response payload */ }
    }
    ```

    ## Authentication
    - Access tokens: Sent in `Authorization: Bearer <token>` header
    - Refresh tokens: Stored in HTTP-only cookies (automatic handling by browser)
    - Token refresh: Use `/auth/refresh` endpoint (no body needed)
servers:
- url: https://api.example.com/api/v1
...
paths:
...
  /users/me:
    get:
      summary: Get current user's profile
      tags:
      - Users
      responses:
        '200':
          description: User fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponseUserFetch'
              examples:
                success:
                  value:
                    meta:
                      code: '200200'
                      type: 'USER_FETCH_SUCCESS'
                      message: 'User fetched successfully'
                      timestamp: '2025-11-04T12:00:00.000Z'
                      request_id: 'aaaaaaaa-bbbb-4ccc-8ddd-aaaaaaaaaaaa'
                      request_duration: 45
                    data:
                      user:
                        id: 'aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa'
                        email: 'storius.test@gmail.com'
                        full_name: 'Storius'
                        avatar: 'https://cdn.example.com/avatars/aaaaaaaa.png'
                        created_at: '2025-10-01T08:00:00.000Z'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
              examples:
                unauthorized:
                  value:
                    meta:
                      code: '401002'
                      type: 'AUTH_UNAUTHORIZED'
                      message: 'Unauthorized access'
                      timestamp: '2025-11-04T12:00:05.000Z'
                      request_id: 'bbbbbbbb-cccc-4ddd-8eee-bbbbbbbbbbbb'
                      request_duration: 2
                    data:
                      details: 'Missing or invalid access token'
                      path: '/api/v1/users/me'
                      method: 'GET'

    patch:
      summary: Update current user's profile
      tags:
      - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
            examples:
              sample:
                value:
                  full_name: 'Storius New'
                  avatar: 'https://cdn.example.com/tmp/aaaaaaaa.png'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponseUserUpdate'
              examples:
                success:
                  value:
                    meta:
                      code: '200201'
                      type: 'USER_UPDATE_SUCCESS'
                      message: 'User updated successfully'
                      timestamp: '2025-11-04T12:05:00.000Z'
                      request_id: 'cccccccc-dddd-4eee-8fff-cccccccccccc'
                      request_duration: 30
                    data:
                      user:
                        id: 'aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa'
                        email: 'storius.test@gmail.com'
                        full_name: 'Storius New'
                        avatar: 'https://cdn.example.com/avatars/aaaaaaaa.png'
                        updated_at: '2025-11-04T12:05:00.000Z'
        '400':
          description: Validation error (bad payload)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
              examples:
                invalid_payload:
                  value:
                    meta:
                      code: '400001'
                      type: 'VALIDATION_ERROR'
                      message: 'Bad request'
                      timestamp: '2025-11-04T12:05:05.000Z'
                      request_id: 'dddddddd-eeee-4fff-8aaa-dddddddddddd'
                      request_duration: 8
                    data:
                      details: 'full_name is required'
                      path: '/api/v1/users/me'
                      method: 'PATCH'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
              examples:
                not_auth:
                  value:
                    meta:
                      code: '401002'
                      type: 'AUTH_UNAUTHORIZED'
                      message: 'Unauthorized access'
                      timestamp: '2025-11-04T12:05:10.000Z'
                      request_id: 'eeeeeeee-ffff-4aaa-8bbb-eeeeeeeeeeee'
                      request_duration: 3
                    data:
                      details: 'Missing or invalid access token'
                      path: '/api/v1/users/me'
                      method: 'PATCH'

    delete:
      summary: Soft delete current user's account
      tags:
      - Users
      responses:
        '200':
          description: User deleted (soft)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponseGenericSuccess'
              examples:
                success:
                  value:
                    meta:
                      code: '200000'
                      type: 'SUCCESS'
                      message: 'Operation completed successfully'
                      timestamp: '2025-11-04T12:10:00.000Z'
                      request_id: 'ffffffff-1111-4bbb-8ccc-ffffffffffff'
                      request_duration: 56
                    data:
                      details: 'User soft-deleted'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
              examples:
                not_auth:
                  value:
                    meta:
                      code: '401002'
                      type: 'AUTH_UNAUTHORIZED'
                      message: 'Unauthorized access'
                      timestamp: '2025-11-04T12:10:05.000Z'
                      request_id: '12121212-1212-4121-8121-121212121212'
                      request_duration: 2
                    data:
                      details: 'Missing or invalid access token'
                      path: '/api/v1/users/me'
                      method: 'DELETE'
  /users/me/avatar:
    post:
      summary: Upload or change user's avatar
      tags:
      - Users
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
            encoding:
              avatar:
                contentType: image/png, image/jpeg
      responses:
        '200':
          description: Avatar upload success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponseAvatarUpload'
              examples:
                success:
                  value:
                    meta:
                      code: '200300'
                      type: 'UPLOAD_FILE_SUCCESS'
                      message: 'File uploaded successfully'
                      timestamp: '2025-11-04T12:15:00.000Z'
                      request_id: '21212121-2121-4212-8121-212121212121'
                      request_duration: 140
                    data:
                      avatar_url: 'https://cdn.example.com/avatars/aaaaaaaa.png'
                      uploaded_at: '2025-11-04T12:15:00.000Z'
        '400':
          description: Invalid file type or too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
              examples:
                invalid_file:
                  value:
                    meta:
                      code: '415001'
                      type: 'UPLOAD_INVALID_TYPE'
                      message: 'Invalid file type'
                      timestamp: '2025-11-04T12:15:05.000Z'
                      request_id: '31313131-3131-4313-8131-313131313131'
                      request_duration: 3
                    data:
                      details: 'File must be PNG or JPEG and <= 2MB'
                      path: '/api/v1/users/me/avatar'
                      method: 'POST'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
              examples:
                not_auth:
                  value:
                    meta:
                      code: '401002'
                      type: 'AUTH_UNAUTHORIZED'
                      message: 'Unauthorized access'
                      timestamp: '2025-11-04T12:15:08.000Z'
                      request_id: '41414141-4141-4414-8141-414141414141'
                      request_duration: 2
                    data:
                      details: 'Missing or invalid access token'
                      path: '/api/v1/users/me/avatar'
                      method: 'POST'
...
components:
  schemas:
...
    Meta:
      type: object
      properties:
        code:
          type: string
          example: '200000'
        type:
          type: string
          example: 'SUCCESS'
        message:
          type: string
          example: 'Operation completed successfully'
        timestamp:
          type: string
          format: date-time
          example: '2025-10-19T10:30:00.000Z'
        request_id:
          type: string
          format: uuid
          example: '11111111-1111-4111-8111-111111111111'
        request_duration:
          type: integer
          example: 123

    # -------------------------
    # CATEGORY SCHEMAS
    # -------------------------
...
    StandardErrorResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: object
          properties:
            details:
              type: string
              example: 'Detailed error description'
            path:
              type: string
              example: '/api/v1/auth/register'
            method:
              type: string
              example: 'POST'
    StandardResponseGenericSuccess:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: object
          properties:
            details:
              type: string
              example: 'Operation completed successfully'

    # -------------------------
    # AUTH SPECIFIC
    # -------------------------
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 'aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa'
        email:
          type: string
          example: 'storius.test@gmail.com'
        fullName:
          type: string
          example: 'Storius'
        avatar:
          type: string
          example: 'https://cdn.example.com/avatars/aaaaaaaa.png'
        createdAt:
          type: string
          format: date-time
          example: '2025-10-01T08:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2025-11-04T12:05:00.000Z'
    StandardResponseUserFetch:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
    StandardResponseUserUpdate:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
    StandardResponseAvatarUpload:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: object
          properties:
            avatar_url:
              type: string
              example: 'https://cdn.example.com/avatars/aaaaaaaa.png'
            uploaded_at:
              type: string
              format: date-time
              example: '2025-11-04T12:15:00.000Z'

    # -------------------------
    # USER SETTINGS
    # -------------------------
...
    UserUpdateRequest:
      type: object
      properties:
        full_name:
          type: string
          example: 'Storius New'
        avatar:
          type: string
          example: 'https://cdn.example.com/tmp/aaaaaaaa.png'
...
  securitySchemes: {}
...